registrando e excluindo chaves Pix no BCB

+++ key-manager-grpc/src/main/kotlin/br/com/zup/pix/remove/RemoveChave.kt
@@ -1,7 +1,10 @@
 package br.com.zup.pix.remove
 
+import br.com.zup.integracao.bcb.BCBPixClient
+import br.com.zup.integracao.bcb.DeletePixKeyRequest
 import br.com.zup.pix.ChavePixRepository
 import br.com.zup.pix.registra.ChavePix
+import io.micronaut.http.HttpStatus
 import io.micronaut.validation.Validated
 import org.slf4j.LoggerFactory
 import java.util.*
@@ -13,7 +16,8 @@ import javax.validation.Valid
 @Singleton
 @Validated
 class RemoveChave(
-    @Inject val chavePixRepository: ChavePixRepository
+    @Inject val chavePixRepository: ChavePixRepository,
+    @Inject val bcbPixClient: BCBPixClient
 ) {
 
     private val logger = LoggerFactory.getLogger(this::class.java)
@@ -29,6 +33,16 @@ class RemoveChave(
             throw ChavePixNaoEncontradaException("Chave Pix não encontrada ou não pertence ao cliente.")
         }
         val chavePix = possivelChavePix.get()
+
+        val bcbResponse = bcbPixClient.removeChave(
+            key = chavePix.chave,
+            request = DeletePixKeyRequest(chavePix.chave)
+        )
+
+        if (bcbResponse.status != HttpStatus.OK) {
+            throw IllegalStateException("Falha ao cadastrar chave no Banco Central")
+        }
+
         chavePixRepository.delete(chavePix)
         logger.info("Chave Pix (${chavePix.id}) removida.")
